Clarify API level in README
It took me ages to work out why somebody would create a compatibility library in 2013 for a feature added to Android in version 4 until I clicked through to the Android documentation and saw API Level 14. I have added the API level to the README to make this more clear.

I'm so used to people saying ICS that "Android 4" hardly means anything to me!